+++ lib/src/auxiliary/mp_segment.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+@immutable
+abstract class MPSegment {
+  const MPSegment({required this.start, required this.end});
+
+  final Offset start;
+  final Offset end;
+}