+++ lib/src/auxiliary/mp_straight_segment.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+import 'package:mapiah/src/auxiliary/mp_segment.dart';
+
+@immutable
+class MPStraightSegment extends MPSegment {
+  const MPStraightSegment({required super.start, required super.end});
+}