+++ lib/src/elements/auxiliary/mp_line_segment_size_orientation_info.dart
@@ -1,15 +1,31 @@
 import 'package:flutter/material.dart';
+import 'package:mapiah/src/auxiliary/mp_numeric_aux.dart';
+import 'package:mapiah/src/constants/mp_constants.dart';
+import 'package:mapiah/src/elements/th_file.dart';
 
 class MPLineSegmentSizeOrientationInfo {
   final int mpID;
   final Offset canvasPosition;
-  final double? size;
-  final double? orientation;
+  late final double size;
+  late final double orientation;
 
   MPLineSegmentSizeOrientationInfo({
     required this.mpID,
     required this.canvasPosition,
-    required this.size,
-    required this.orientation,
-  });
+    required double? size,
+    required double? orientation,
+    required THFile thFile,
+  }) {
+    this.size = (size == null) ? mpSlopeLinePointDefaultLSize : size;
+
+    if (orientation == null) {
+      final Offset tangent = MPNumericAux.segmentTangent(mpID, thFile);
+      final Offset normal = MPNumericAux.normalFromTangent(tangent);
+      final double azimuth = MPNumericAux.directionOffsetToDegrees(normal);
+
+      this.orientation = azimuth;
+    } else {
+      this.orientation = orientation;
+    }
+  }
 }