+++ lib/src/widgets/inputs/mp_azimuth_picker_widget.dart
@@ -1,11 +1,9 @@
 import 'dart:math' as math;
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:mapiah/main.dart';
 import 'package:mapiah/src/auxiliary/mp_numeric_aux.dart';
 import 'package:mapiah/src/constants/mp_constants.dart';
-import 'package:mapiah/src/generated/i18n/app_localizations.dart';
+import 'package:mapiah/src/painters/mp_compass_painter.dart';
 
 class MPAzimuthPickerWidget extends StatefulWidget {
   final double initialAzimuth;
@@ -77,14 +75,14 @@ class _MPAzimuthPickerWidgetState extends State<MPAzimuthPickerWidget> {
   double _calculateAngle({required Offset center, required Offset position}) {
     final Offset delta = position - center;
     // Convert cartesian to polar coordinates (with y inverted)
-    final double angle = math.atan2(delta.dx, -delta.dy) * 180 / math.pi;
+    final double angle = math.atan2(delta.dx, -delta.dy) * mp1Radian;
 
     return angle;
   }
 
   @override
   Widget build(BuildContext context) {
-    final double compassSize = widget.size * 0.8;
+    final double compassBoxSize = widget.size * mpCompassBoxSizeFactor;
 
     return Column(
       mainAxisSize: MainAxisSize.min,
@@ -92,26 +90,27 @@ class _MPAzimuthPickerWidgetState extends State<MPAzimuthPickerWidget> {
         // Compass circle with arrow
         GestureDetector(
           onPanUpdate: (details) {
-            _updateAzimuthOnTap(details.globalPosition, compassSize);
+            _updateAzimuthOnTap(details.globalPosition, compassBoxSize);
           },
           onTapDown: (details) {
-            _updateAzimuthOnTap(details.globalPosition, compassSize);
+            _updateAzimuthOnTap(details.globalPosition, compassBoxSize);
           },
           child: SizedBox(
-            width: compassSize,
-            height: compassSize,
+            width: compassBoxSize,
+            height: compassBoxSize,
             child: CustomPaint(
-              painter: _CompassPainter(
+              painter: MPCompassPainter(
                 azimuth: _azimuth,
-                markerSize: _markerSize,
+                arrowLength: _markerSize,
+                drawBackgroundLines: true,
               ),
             ),
           ),
         ),
-        const SizedBox(height: 16),
+        const SizedBox(height: mpButtonSpace * 2),
         // Text input field
         SizedBox(
-          width: compassSize * 0.6,
+          width: compassBoxSize * 0.6,
           child: TextField(
             controller: _controller,
             focusNode: widget.focusNode,
@@ -155,167 +154,5 @@ class _MPAzimuthPickerWidgetState extends State<MPAzimuthPickerWidget> {
         : angle;
 
     _updateAzimuth(adjustedAngle, updateTextField: true);
-  }
-}
-
-class _CompassPainter extends CustomPainter {
-  final double azimuth;
-  final double markerSize;
-
-  _CompassPainter({required this.azimuth, required this.markerSize});
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final AppLocalizations appLocalizations = mpLocator.appLocalizations;
-    final Offset center = Offset(size.width, size.height) / 2;
-    final double radius = size.width / 2;
-
-    // Draw compass circle
-    final Paint circlePaint = Paint()
-      ..color = Colors.grey[200]!
-      ..style = PaintingStyle.fill;
-
-    canvas.drawCircle(center, radius, circlePaint);
-
-    // Draw border
-    final Paint borderPaint = Paint()
-      ..color = Colors.grey
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 2;
-
-    canvas.drawCircle(center, radius, borderPaint);
-
-    // Draw background lines
-    _drawBackgroundLines(canvas, center, radius);
-
-    // Draw cardinal directions
-    final TextPainter textPainter = TextPainter(
-      textAlign: TextAlign.center,
-      textDirection: TextDirection.ltr,
-    );
-
-    final List<String> directions = [
-      appLocalizations.mpAzimuthNorthAbbreviation,
-      appLocalizations.mpAzimuthEastAbbreviation,
-      appLocalizations.mpAzimuthSouthAbbreviation,
-      appLocalizations.mpAzimuthWestAbbreviation,
-    ];
-
-    const List<double> rightAngles = [0.0, 90.0, 180.0, 270.0];
-
-    for (int i = 0; i < directions.length; i++) {
-      final double angleRad = rightAngles[i] * math.pi / 180;
-      final Offset textOffset = Offset(
-        center.dx + math.sin(angleRad) * radius * 0.83,
-        center.dy - math.cos(angleRad) * radius * 0.83,
-      );
-
-      textPainter.text = TextSpan(
-        text: directions[i],
-        style: TextStyle(
-          color: Colors.black,
-          fontSize: markerSize * 0.7,
-          fontWeight: FontWeight.bold,
-        ),
-      );
-      textPainter.layout();
-      textPainter.paint(
-        canvas,
-        textOffset - Offset(textPainter.width / 2, textPainter.height / 2),
-      );
-    }
-
-    // Save the canvas state
-    canvas.save();
-
-    // Translate the canvas to the center of the compass
-    canvas.translate(center.dx, center.dy);
-
-    // Rotate the canvas based on the azimuth
-    canvas.rotate(azimuth * math.pi / 180);
-
-    // Draw the arrow body
-    final Paint arrowBodyPaint = Paint()
-      ..color = Colors.red
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 3;
-
-    final double arrowLength = radius * 0.6;
-    final Offset arrowTipBase = Offset(0, -arrowLength * 0.9);
-
-    canvas.drawLine(Offset.zero, arrowTipBase, arrowBodyPaint);
-
-    // Draw the arrowhead
-    final Paint arrowPaint = Paint()
-      ..color = Colors.red
-      ..style = PaintingStyle.fill;
-
-    final Offset arrowTip = Offset(0, -arrowLength);
-    final double arrowSide = markerSize * 0.4;
-    final double arrowBaseLength = -arrowLength * 0.78;
-    final Offset arrowSide1 = Offset(-arrowSide, arrowBaseLength);
-    final Offset arrowSide2 = Offset(arrowSide, arrowBaseLength);
-
-    final Path path = Path();
-
-    path.moveTo(arrowTip.dx, arrowTip.dy);
-    path.lineTo(arrowSide1.dx, arrowSide1.dy);
-    path.lineTo(arrowTipBase.dx, arrowTipBase.dy);
-    path.lineTo(arrowSide2.dx, arrowSide2.dy);
-    path.close();
-
-    canvas.drawPath(path, arrowPaint);
-
-    // Draw central circle
-    canvas.drawCircle(Offset.zero, markerSize * 0.2, arrowPaint);
-
-    // Restore the canvas state
-    canvas.restore();
-  }
-
-  void _drawBackgroundLines(Canvas canvas, Offset center, double radius) {
-    // Paint for the lines
-    final Paint rightAnglePaint = Paint()
-      ..color = Colors.grey.shade700
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 1.5;
-    final Paint fortyFiveDegreePaint = Paint()
-      ..color = Colors.grey
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 1;
-
-    /// radius90 is smaller than radius45 to make right angle lines longer so
-    /// there is space to write the cardinal direction letters.
-    final double radius90 = radius * 0.68;
-    final double radius45 = radius * 0.87;
-
-    for (int i = 0; i < 4; i++) {
-      final double angleRad90 = i * math.pi / 2;
-      final double angleRad45 = angleRad90 + (math.pi / 4);
-
-      // Draw right angle lines (0°, 90°, 180°, 270°)
-      final Offset lineEnd90 = Offset(
-        center.dx + math.cos(angleRad90) * radius90,
-        center.dy + math.sin(angleRad90) * radius90,
-      );
-
-      canvas.drawLine(center, lineEnd90, rightAnglePaint);
-
-      // Draw 45° lines (45°, 135°, 225°, 315°)
-      final Offset lineEnd45 = Offset(
-        center.dx + math.cos(angleRad45) * radius45,
-        center.dy + math.sin(angleRad45) * radius45,
-      );
-
-      canvas.drawLine(center, lineEnd45, fortyFiveDegreePaint);
-    }
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    if (identical(this, oldDelegate)) return false;
-
-    return azimuth != (oldDelegate as _CompassPainter).azimuth ||
-        markerSize != oldDelegate.markerSize;
   }
 }